InfoProxyItemSearch -> InfoProxyCatalogSearch; InfoProxy11 -> InfoProxyItemSearch
These were misnamed when they were added.
